JuliaBoyz

Documentation for JuliaBoyz.

JuliaBoyz.check_solution

JuliaBoyz.generate_adjacency_list

JuliaBoyz.generate_adjacency_matrix

JuliaBoyz.generate_upper_bound

JuliaBoyz.get_adjacent_streets

JuliaBoyz.greedy_dfs_solution

JuliaBoyz.read_city

JuliaBoyz.shortest_path

JuliaBoyz.solution_distance

JuliaBoyz.check_solution — Method

Given a city and a potential solution, checks the validity of the solution.

Works by beginning at the first junction in an itinerary and check if the next junction is in the adjacency list of the current node, returning false if it is not, and continuing on otherwise. Repeat until the end of the itinerary, and do this for all itineraries in the potential solution.

JuliaBoyz.generate_adjacency_list — Method

Generates an adjacency list for the junctions of the currently loaded city to represent the graph. Returns a Dict mapping each junction index to a Set{Int64} containing all adjacent junctions' indices.

JuliaBoyz.generate_adjacency_matrix — Method

Generates an adjacency matrix for the junctions of the currently loaded city to represent the graph. Returns a Tuple of two sparse matrices containing all the durations and distances, respectively, of streets between a starting junction (rows) and an ending junction (columns).

JuliaBoyz.generate_upper_bound — Method

Naively generates an upper bound on the possible distance by sorting all available streets by duration to distance ratio, and adding until all time is used up (across all cars), or all streets are traversed.

JuliaBoyz.get_adjacent_streets — Method

Generates an adjacency list for the streets adjacent to all junctions of the currently loaded city to represent the graph. Returns a Dict mapping each junction index to a Vector{Street} containing all its adjacent streets.

JuliaBoyz.greedy_dfs_solution — Method

ADD DOCSTRING

JuliaBoyz.read_city — Method

Given the path for a file containing a text file of a problem in the specified format, creates a JBCity instance with the relevant information.

JuliaBoyz.shortest_path — Method

Computes the shortest path between two given junctions using Djikstra's algorithm. Returns a Tuple containing the sequence of junctions (represented by their indices, including the start/finish junctions) that give the shortest path and the duration of the path.

JuliaBoyz.solution_distance — Method

solution_distance(sol)

Computes the distance covered by all 8 cars.