JuliaBoyz

Documentation for JuliaBoyz.

JuliaBoyz.generate_adjacency_list

JuliaBoyz.generate_adjacency_matrix

JuliaBoyz.generate_greedy_random_solution

JuliaBoyz.generate_upper_bound

JuliaBoyz.get_adjacent_streets

JuliaBoyz.generate_adjacency_list — Method

Generates an adjacency list for the junctions of the currently loaded city to represent the graph.

JuliaBoyz.generate_adjacency_matrix — Method

Generates an adjacency matrix for the junctions of the currently loaded city to represent the graph.

JuliaBoyz.generate_greedy_random_solution — Method

Uses a greedy algorithm to generate itineraries for the cars in the currently loaded city.

For each car, we look at the adjacency lists of the current node and choose the one with the highest distance to duration ratio, which is done by choosing the first street in the adjacency list of the current node. After traversing a street, add it to a list of already traversed streets. If a street has already been traversed, we skip it, and go to the next one in the adjacency list. If all adjacent streets from a node are already traveresd, choose a random one.

JuliaBoyz.generate_upper_bound — Method

Naively generates an upper bound on the possible distance by sorting all available streets by duration to distance ratio, and adding until all time is used up, or all streets are traversed.

JuliaBoyz.get_adjacent_streets — Method

Perhaps best of both worlds? Stores the street, and therefore, both the adjacent junctions and the properties of the street.